Fastq_sample: simplify representation
Fastq_sample.t now represents a single single-end or paired-end fastq
file. Bowtie2 accepts a single fastq_sample as argument and possibly a
list of additional files.

Note that this is not a limitation in the sense that if a given
biological sample is scattered in several fastq files (for instance
several sequencing lanes), the results can be gathered at alignment
time, or peak calling time or feature count time.